Add `PostProcessor` inline resource and support Kyverno during post-processing (#34)
Adds the `inline.nyl.io/v1/PostProcessor` resource that can be used to
apply Kyverno policies to the resources defined in a Manifest. See
documentation for a usage example.

I also added a Glossary page to the documentation (which we are not yet
fully compatible with in the code but it lays out the intended meaning
of the words, e.g. "manifest" is a file that contains many resources,
but we currently use "manifest" as an analog for "resource" in the
code).